Apply patch from Evgenii Philippov <efilippov@gmail.com>:
I renamed DocSplitter to DocSplitterOpenNLPImpl, created a DocSplitter
interface which is obtained by DocSplitterFactory.
DocSplitterFallbackImpl is used when OpenNLP lib is not present.

One line is changed in RelationExtractor to call the factory.

The trick is to NOT reference the opennlp classes in the class code,
but you referenced opennlp.SentenceDetector in your code, that's why
it was failing when the opennlp lib was not present.

Signed-off-by: Evgenii Philippov <efilippov@gmail.com>