Fix for bug #236
The "obvious" fix is to remove the check for the ref node. Its
not clear why that check was ever there. the "ref" node should
almost surely always be ther, since the algs files creat this for
almost all situations, so its kind of strange that no alg created
a ref for "going"  (test sentence: "Tom asks Roy how they are going
to get the gold") so maybe the algs are broken.

Whatever. This fix works, and its safe. (Should be safe!?)